+++ resourcecode/eva/__init__.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 # Extreme Values Modelling
 
-# copyright 2021 IFREMER (Brest, FRANCE), all rights reserved.
+# Copyright 2020-2022 IFREMER (Brest, FRANCE), all rights reserved.
 # contact -- mailto:nicolas.raillard@ifremer.fr
 #
 # This file is part of Resourcecode.